Add preconfigured server.properties

+++ launch.sh
@@ -26,6 +26,13 @@ if [ ! -f "/minecraft/${SERVER}" ]; then
         echo "eula=$EULA" >> eula.txt
     fi
 
+    if [ ! -f "/minecraft/server.properties" ]; then
+        export MOTD=$MOTD
+        export RCON_PW=$RCON_PW
+
+        envsubst "$MOTD:$RCONPW /mc/server.properties > /minecraft/server.properties
+    fi
+
     java -jar "/tmp/mc/forge-${MC_VERSION}-${FORGE_VERSION}-installer.jar" --installServer
 
     mkdir -p /minecraft/mods
@@ -35,21 +42,26 @@ if [ ! -f "/minecraft/${SERVER}" ]; then
     rm -r /tmp/mc
 fi
 
-if [ ! -f "/minecraft/mods/${SPONGE}" ]; then
-    mkdir -p /tmp/mc
-
-    cd /tmp/mc
-    
-    echo "Downloading sponge ${SPONGE_DOWNLOAD_URL}"
-    curl -s "${SPONGE_DOWNLOAD_URL}" -O
-
-    rm -f "/minecraft/mods/spongeforge-*.jar"
-
-    mv "/tmp/mc/${SPONGE}" /minecraft/mods/
-
-    cd /minecraft
-
-    rm -r /tmp/mc
+if [ ${FORGE_VERSION} = "latest" ]; then
+  curl -fsSL -o /tmp/forge.json http://files.minecraftforge.net/maven/net/minecraftforge/forge/promotions_slim.json
+  FORGE_VERSION=$(cat /tmp/forge.json | jq -r ".promos[\"$MC_VERSION-latest\"]")
+else
+  if [ ! -f "/minecraft/mods/${SPONGE}" ]; then
+      mkdir -p /tmp/mc
+  
+      cd /tmp/mc
+      
+      echo "Downloading sponge ${SPONGE_DOWNLOAD_URL}"
+      curl -s "${SPONGE_DOWNLOAD_URL}" -O
+  
+      rm -f "/minecraft/mods/spongeforge-*.jar"
+  
+      mv "/tmp/mc/${SPONGE}" /minecraft/mods/
+  
+      cd /minecraft
+  
+      rm -r /tmp/mc
+  fi
 fi
 
 cd /minecraft
